requirements.txt
fastapi
uvicorn
python-dotenv
requests
paddleocr
paddlepaddle==2.5.0
